Make enum types appearing as struct members be as small as possible.
Bug: 420965163

+++ core/fxge/fx_fontencoding.h
@@ -9,7 +9,7 @@
 
 namespace fxge {
 
-enum class FontEncoding : uint32_t {
+enum class FontEncoding : uint8_t {
   kAdobeCustom,
   kAdobeExpert,
   kAdobeStandard,
